Remove testcontainers database module and related references from documentation and code
