+++ test/native/src/test/java/org/apache/shardingsphere/test/natived/testcontainers/TcPostgreSQLDatabaseType.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.database.testcontainers.type;
+package org.apache.shardingsphere.test.natived.testcontainers;
 
 import org.apache.shardingsphere.infra.database.core.type.DatabaseType;
 import org.apache.shardingsphere.infra.spi.type.typed.TypedSPILoader;
@@ -25,22 +25,22 @@ import java.util.Collections;
 import java.util.Optional;
 
 /**
- * Database type of MariaDB in testcontainers-java.
+ * Database type of PostgreSQL in testcontainers.
  */
-public final class TcMariaDBDatabaseType implements TestcontainersDatabaseType {
+public final class TcPostgreSQLDatabaseType implements TestcontainersDatabaseType {
     
     @Override
     public Collection<String> getJdbcUrlPrefixes() {
-        return Collections.singleton("jdbc:tc:mariadb:");
+        return Collections.singleton("jdbc:tc:postgresql:");
     }
     
     @Override
     public Optional<DatabaseType> getTrunkDatabaseType() {
-        return Optional.of(TypedSPILoader.getService(DatabaseType.class, "MariaDB"));
+        return Optional.of(TypedSPILoader.getService(DatabaseType.class, "PostgreSQL"));
     }
     
     @Override
     public String getType() {
-        return "TC-MariaDB";
+        return "TC-PostgreSQL";
     }
 }
