+++ test/native/src/test/resources/META-INF/services/org.apache.shardingsphere.infra.database.core.type.DatabaseType
@@ -15,11 +15,5 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.infra.database.testcontainers.type.TcClickHouseDatabaseType
-org.apache.shardingsphere.infra.database.testcontainers.type.TcMariaDBDatabaseType
-org.apache.shardingsphere.infra.database.testcontainers.type.TcMySQLDatabaseType
-org.apache.shardingsphere.infra.database.testcontainers.type.TcOracleDatabaseType
-org.apache.shardingsphere.infra.database.testcontainers.type.TcPostgreSQLDatabaseType
-org.apache.shardingsphere.infra.database.testcontainers.type.TcSQLServerDatabaseType
-org.apache.shardingsphere.infra.database.testcontainers.type.TcTiDBDatabaseType
-org.apache.shardingsphere.infra.database.testcontainers.type.TcFirebirdDatabaseType
+org.apache.shardingsphere.test.natived.testcontainers.TcPostgreSQLDatabaseType
+org.apache.shardingsphere.test.natived.testcontainers.TcSQLServerDatabaseType
